requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
openpyxl==3.1.2
xlrd==2.0.1
PyPDF2==3.0.1
python-docx==1.0.1
scikit-learn==1.3.1
wordcloud==1.9.2
textstat==0.7.3
nltk==3.8.1
pillow==10.2.0
statsmodels==0.14.0